@@ -6,7 +6,7 @@ description: Demonstrates how to use the twilio_conversations plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -21,10 +21,10 @@ dependencies:
     path: ../
 
   cupertino_icons: ^1.0.0
-  flutter_riverpod: ^0.12.1
-  intl: ^0.16.1
-  mime_type: ^0.3.2
-
+  provider: ^6.0.0
+  http: ^0.13.3
+  intl: ^0.17.0
+  mime_type: ^1.0.0
 
 dev_dependencies:
   flutter_test:
