@@ -6,17 +6,15 @@ repository: https://gitlab.com/twilio-flutter/conversations
 issue_tracker: https://gitlab.com/twilio-flutter/conversations/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
   build_runner: ^1.10.9
-  enum_to_string: ^1.0.14
-  json_annotation: ^3.1.1
-  json_serializable: ^3.5.1
-  mime_type: ^0.3.2
+  enum_to_string: ^2.0.1
+  mime_type: ^1.0.0
 
 dev_dependencies:
   flutter_test:
