@@ -1,23 +1,26 @@
 import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:enum_to_string/enum_to_string.dart';
 import 'package:flutter/services.dart';
+import 'package:mime_type/mime_type.dart';
 import 'package:twilio_conversations/twilio_conversations.dart';
 
 class Conversation {
   final String sid;
-  Attributes attributes;
-  String uniqueName;
-  String friendlyName;
-  ConversationStatus status;
-  ConversationSynchronizationStatus synchronizationStatus;
-  DateTime dateCreated;
-  String createdBy;
-  DateTime dateUpdated;
-  DateTime lastMessageDate;
-  int lastMessageIndex;
-  int lastReadMessageIndex;
+  Attributes? attributes;
+  String? uniqueName;
+  String? friendlyName;
+  ConversationStatus status = ConversationStatus.UNKNOWN;
+  ConversationSynchronizationStatus synchronizationStatus =
+      ConversationSynchronizationStatus.NONE;
+  DateTime? dateCreated;
+  String? createdBy;
+  DateTime? dateUpdated;
+  DateTime? lastMessageDate;
+  int? lastMessageIndex;
+  int? lastReadMessageIndex;
 
   bool get hasMessages => lastMessageIndex != null;
 
@@ -26,17 +29,14 @@ class Conversation {
           synchronizationStatus == ConversationSynchronizationStatus.ALL) ||
       (status == ConversationStatus.NOT_PARTICIPATING &&
           synchronizationStatus == ConversationSynchronizationStatus.METADATA);
-  Messages messages;
-
-  Participants participants;
 
   /// Local caching event stream so each instance will use the same stream.
-  static final Map<String, Stream> _conversationStreams = {};
   final StreamController<Message> _onMessageAddedCtrl =
       StreamController<Message>.broadcast();
 
   /// Called when a [Message] is added to the conversation the current user is subscribed to.
-  Stream<Message> onMessageAdded;
+  late Stream<Message> onMessageAdded;
+
   final StreamController<MessageUpdatedEvent> _onMessageUpdatedCtrl =
       StreamController<MessageUpdatedEvent>.broadcast();
 
@@ -44,7 +44,7 @@ class Conversation {
   ///
   /// You could obtain the [Conversation] where it was updated by using [Message.getConversation] or [Message.conversationSid].
   /// [Message] change events include body updates and attribute updates.
-  Stream<MessageUpdatedEvent> onMessageUpdated;
+  late Stream<MessageUpdatedEvent> onMessageUpdated;
 
   final StreamController<Message> _onMessageDeletedCtrl =
       StreamController<Message>.broadcast();
@@ -52,7 +52,7 @@ class Conversation {
   /// Called when a [Message] is deleted from the conversation the current user is subscribed to.
   ///
   /// You could obtain the [Conversation] where it was deleted by using [Message.getConversation] or [Message.conversationSid].
-  Stream<Message> onMessageDeleted;
+  late Stream<Message> onMessageDeleted;
 
   final StreamController<Participant> _onParticipantAddedCtrl =
       StreamController<Participant>.broadcast();
@@ -60,7 +60,7 @@ class Conversation {
   /// Called when a [Participant] is added to the conversation the current user is subscribed to.
   ///
   /// You could obtain the [Conversation] where it was added by using [Participant.getConversation].
-  Stream<Participant> onParticipantAdded;
+  late Stream<Participant> onParticipantAdded;
 
   final StreamController<ParticipantUpdatedEvent> _onParticipantUpdatedCtrl =
       StreamController<ParticipantUpdatedEvent>.broadcast();
@@ -69,7 +69,7 @@ class Conversation {
   ///
   /// You could obtain the [Conversation] where it was updated by using [Participant.getConversation].
   /// [Participant] change events include body updates and attribute updates.
-  Stream<ParticipantUpdatedEvent> onParticipantUpdated;
+  late Stream<ParticipantUpdatedEvent> onParticipantUpdated;
 
   final StreamController<Participant> _onParticipantDeletedCtrl =
       StreamController<Participant>.broadcast();
@@ -77,14 +77,14 @@ class Conversation {
   /// Called when a [Participant] is deleted from the conversation the current user is subscribed to.
   ///
   /// You could obtain the [Conversation] where it was deleted by using [Participant.getConversation].
-  Stream<Participant> onParticipantDeleted;
+  late Stream<Participant> onParticipantDeleted;
 
   //#region Typing events
   final StreamController<TypingEvent> _onTypingStartedCtrl =
       StreamController<TypingEvent>.broadcast();
 
   /// Called when an [Participant] starts typing in a [Conversation].
-  Stream<TypingEvent> onTypingStarted;
+  late Stream<TypingEvent> onTypingStarted;
 
   final StreamController<TypingEvent> _onTypingEndedCtrl =
       StreamController<TypingEvent>.broadcast();
@@ -92,17 +92,15 @@ class Conversation {
   /// Called when an [Participant] stops typing in a [Conversation].
   ///
   /// Typing indicator has a timeout after user stops typing to avoid triggering it too often. Expect about 5 seconds delay between stopping typing and receiving typing ended event.
-  Stream<TypingEvent> onTypingEnded;
+  late Stream<TypingEvent> onTypingEnded;
 
   final StreamController<Conversation> _onSynchronizationChangedCtrl =
       StreamController<Conversation>.broadcast();
 
   /// Called when conversation synchronization status changed.
-  Stream<Conversation> onSynchronizationChanged;
+  late Stream<Conversation> onSynchronizationChanged;
 
   Conversation(this.sid) {
-    this.messages = Messages(this);
-    this.participants = Participants(conversationSid: sid);
     onMessageAdded = _onMessageAddedCtrl.stream;
     onMessageUpdated = _onMessageUpdatedCtrl.stream;
     onMessageDeleted = _onMessageDeletedCtrl.stream;
@@ -112,35 +110,36 @@ class Conversation {
     onTypingStarted = _onTypingStartedCtrl.stream;
     onTypingEnded = _onTypingEndedCtrl.stream;
     onSynchronizationChanged = _onSynchronizationChangedCtrl.stream;
-
-    _conversationStreams[sid] ??=
-        EventChannel('twilio_conversations/$sid').receiveBroadcastStream(0);
-    _conversationStreams[sid].listen(_parseEvents);
   }
 
   void updateFromMap(Map<String, dynamic> map) {
     attributes = map['attributes'] == null
         ? null
-        : Attributes.fromJson(map['attributes'] as Map<String, dynamic>);
-    uniqueName = map['uniqueName'] as String;
-    friendlyName = map['friendlyName'] as String;
-    status = EnumToString.fromString(
-        ConversationStatus.values, map['status'] as String);
+        : Attributes.fromMap(Map<String, dynamic>.from(map['attributes']));
+    uniqueName = map['uniqueName'] as String?;
+    friendlyName = map['friendlyName'] as String?;
+
+    status =
+        EnumToString.fromString(ConversationStatus.values, map['status']) ??
+            ConversationStatus.UNKNOWN;
+
     synchronizationStatus = EnumToString.fromString(
-        ConversationSynchronizationStatus.values,
-        map['synchronizationStatus'] as String);
+            ConversationSynchronizationStatus.values,
+            map['synchronizationStatus']) ??
+        ConversationSynchronizationStatus.NONE;
+
     dateCreated = map['dateCreated'] == null
         ? null
         : DateTime.parse(map['dateCreated'] as String);
-    createdBy = map['createdBy'] as String;
+    createdBy = map['createdBy'] as String?;
     dateUpdated = map['dateUpdated'] == null
         ? null
         : DateTime.parse(map['dateUpdated'] as String);
     lastMessageDate = map['lastMessageDate'] == null
         ? null
         : DateTime.parse(map['lastMessageDate'] as String);
-    lastReadMessageIndex = map['lastReadMessageIndex'] as int;
-    lastMessageIndex = map['lastMessageIndex'] as int;
+    lastReadMessageIndex = map['lastReadMessageIndex'] as int?;
+    lastMessageIndex = map['lastMessageIndex'] as int?;
   }
 
   /// Construct from a map.
@@ -152,34 +151,169 @@ class Conversation {
     return conversation;
   }
 
-  Future<int> getUnreadMessagesCount() async {
+  Future<bool?> addParticipantByIdentity(String identity) async {
+    final result = await TwilioConversations.methodChannel.invokeMethod<bool>(
+        'ParticipantsMethods.addParticipantByIdentity',
+        {'identity': identity, 'conversationSid': sid});
+
+    return result;
+  }
+
+  Future<bool?> removeParticipantByIdentity(String identity) async {
+    final result = await TwilioConversations.methodChannel.invokeMethod<bool>(
+        'ParticipantsMethods.removeParticipantByIdentity',
+        {'identity': identity, 'conversationSid': sid});
+
+    return result;
+  }
+
+  Future<List<Participant>> getParticipantsList() async {
+    final result = await TwilioConversations.methodChannel.invokeMethod(
+        'ParticipantsMethods.getParticipantsList', {'conversationSid': sid});
+
+    var participants = List.from(result)
+        .map((e) => Participant.fromMap(Map<String, dynamic>.from(e)))
+        .toList();
+    return participants;
+  }
+
+  Future<List<User>> getUsers() async {
+    final result = await TwilioConversations.methodChannel
+        .invokeMethod('ParticipantsMethods.getUsers', {'conversationSid': sid});
+
+    var users =
+        result.map<User>((e) => User.fromMap(Map<String, dynamic>.from(e))).toList();
+
+    return users;
+  }
+
+  Future<int?> getUnreadMessagesCount() async {
     final result = await TwilioConversations.methodChannel.invokeMethod<int>(
         'ConversationMethods.getUnreadMessagesCount', {'conversationSid': sid});
 
     return result;
   }
 
-  Future<bool> join() async {
+  Future<bool?> join() async {
     final result = await TwilioConversations.methodChannel.invokeMethod<bool>(
         'ConversationMethods.join', {'conversationSid': sid});
 
     return result;
   }
 
-  Future<bool> leave() async {
+  Future<bool?> leave() async {
     final result = await TwilioConversations.methodChannel.invokeMethod<bool>(
         'ConversationMethods.leave', {'conversationSid': sid});
 
     return result;
   }
 
+  /// Fetch at most count messages including and prior to the specified index.
+  Future<List<Message>> getMessagesBefore({
+    required int index,
+    required int count,
+  }) async {
+    if (!hasMessages) {
+      return [];
+    }
+    try {
+      final result = await TwilioConversations.methodChannel
+          .invokeMethod('MessagesMethods.getMessagesBefore', {
+        'index': index,
+        'count': count,
+        'conversationSid': sid,
+      });
+
+      final messages = result
+          .map<Message>((i) => Message.fromMap(Map<String, dynamic>.from(i)))
+          .toList();
+
+      return messages;
+    } on PlatformException catch (err) {
+      throw TwilioConversations.convertException(err);
+    }
+  }
+
+  Future<List<Message>> getLastMessages(int count) async {
+    if (!hasMessages) {
+      return [];
+    }
+    final result = await TwilioConversations.methodChannel
+        .invokeMethod('MessagesMethods.getLastMessages', {
+      'count': count,
+      'conversationSid': sid,
+    });
+
+    var messages = result
+        .map((e) => Message.fromMap(Map<String, dynamic>.from(e)))
+        .toList();
+    return messages;
+  }
+
+  Future<Message> sendMessage(String messageBody) async {
+    final messageOptions = MessageOptions()..withBody(messageBody);
+
+    final result = await TwilioConversations.methodChannel
+        .invokeMethod('MessagesMethods.sendMessage', {
+      'options': messageOptions.toMap(),
+      'conversationSid': sid,
+    });
+
+    return Message.fromMap(Map<String, dynamic>.from(result));
+  }
+
+  Future<Message> sendAttachment(File attachment) async {
+    final path = mime(attachment.path);
+
+    var messageOptions = MessageOptions();
+    if (path != null) {
+      messageOptions.withMedia(attachment, path);
+    }
+
+    final result = await TwilioConversations.methodChannel
+        .invokeMethod('MessagesMethods.sendMessage', {
+      'options': messageOptions.toMap(),
+      'conversationSid': sid,
+    });
+
+    return Message.fromMap(Map<String, dynamic>.from(result));
+  }
+
+  Future<int> setAllMessagesRead() async {
+    if (!hasMessages) {
+      return 0;
+    }
+    try {
+      return await TwilioConversations.methodChannel
+          .invokeMethod('MessagesMethods.setAllMessagesRead', {
+        'conversationSid': sid,
+      }) as int;
+    } on PlatformException catch (err) {
+      throw TwilioConversations.convertException(err);
+    }
+  }
+
   Future<String> setFriendlyName(String friendlyName) async {
-    final result = await TwilioConversations.methodChannel.invokeMethod(
-        'ConversationMethods.setFriendlyName',
-        {'conversationSid': sid, 'friendlyName': friendlyName});
+    final result = await TwilioConversations.methodChannel
+        .invokeMethod('ConversationMethods.setFriendlyName', {
+      'conversationSid': sid,
+      'friendlyName': friendlyName,
+    });
 
     this.friendlyName = result.toString();
     return friendlyName;
+  }
+
+  Future<int> setLastReadMessageIndex(int lastReadMessageIndex) async {
+    try {
+      return await TwilioConversations.methodChannel.invokeMethod(
+          'MessagesMethods.setLastReadMessageIndex', {
+        'conversationSid': sid,
+        'lastReadMessageIndex': lastReadMessageIndex
+      }) as int;
+    } on PlatformException catch (err) {
+      throw TwilioConversations.convertException(err);
+    }
   }
 
   /// Indicate that Participant is typing in this conversation.
@@ -187,7 +321,7 @@ class Conversation {
   /// You should call this method to indicate that a local user is entering a message into current conversation. The typing state is forwarded to users subscribed to this conversation through [Conversation.onTypingStarted] and [Conversation.onTypingEnded] callbacks.
   /// After approximately 5 seconds after the last [Conversation.typing] call the SDK will emit [Conversation.onTypingEnded] signal.
   /// One common way to implement this indicator is to call [Conversation.typing] repeatedly in response to key input events.
-  Future<bool> typing() async {
+  Future<void> typing() async {
     try {
       return await TwilioConversations.methodChannel
           .invokeMethod('ConversationMethods.typing', {'conversationSid': sid});
@@ -197,26 +331,25 @@ class Conversation {
   }
 
   /// Parse native channel events to the right event streams.
-  void _parseEvents(dynamic event) {
-    var eventMap = jsonDecode(event.toString());
-    final String eventName = eventMap['name'] as String;
-    final data = Map<String, dynamic>.from(eventMap['data'] as Map);
+  void parseEvents(dynamic event) {
+    final eventName = event['name'] as String;
+    final data = Map<String, dynamic>.from(event['data'] as Map);
     if (data['conversation'] != null) {
       final conversationMap =
           Map<String, dynamic>.from(data['conversation'] as Map);
       updateFromMap(conversationMap);
     }
 
-    Message message;
+    Message? message;
     if (data['message'] != null) {
       final messageMap = Map<String, dynamic>.from(data['message'] as Map);
-      message = Message.fromJson(messageMap);
+      message = Message.fromMap(messageMap);
     }
 
-    Participant participant;
+    Participant? participant;
     if (data['participant'] != null) {
       final memberMap = Map<String, dynamic>.from(data['participant'] as Map);
-      participant = Participant.fromJson(memberMap);
+      participant = Participant.fromMap(memberMap);
     }
 
     dynamic reason;
@@ -237,40 +370,46 @@ class Conversation {
 
     switch (eventName) {
       case 'messageAdded':
-        assert(message != null);
-        _onMessageAddedCtrl.add(message);
+        if (message != null) {
+          _onMessageAddedCtrl.add(message);
+        }
         break;
       case 'messageUpdated':
-        assert(message != null);
-        assert(reason != null);
-        _onMessageUpdatedCtrl
-            .add(MessageUpdatedEvent(message, reason as MessageUpdateReason));
+        if (message != null && reason != null) {
+          _onMessageUpdatedCtrl
+              .add(MessageUpdatedEvent(message, reason as MessageUpdateReason));
+        }
         break;
       case 'messageDeleted':
-        assert(message != null);
-        _onMessageDeletedCtrl.add(message);
+        if (message != null) {
+          _onMessageDeletedCtrl.add(message);
+        }
         break;
       case 'participantAdded':
-        assert(participant != null);
-        _onParticipantAddedCtrl.add(participant);
+        if (participant != null) {
+          _onParticipantAddedCtrl.add(participant);
+        }
         break;
       case 'participantUpdated':
-        assert(participant != null);
-        assert(reason != null);
-        _onParticipantUpdatedCtrl.add(ParticipantUpdatedEvent(
-            participant, reason as ParticipantUpdateReason));
+        if (participant != null && reason != null) {
+          _onParticipantUpdatedCtrl.add(ParticipantUpdatedEvent(
+              participant, reason as ParticipantUpdateReason));
+        }
         break;
       case 'participantDeleted':
-        assert(participant != null);
-        _onParticipantDeletedCtrl.add(participant);
+        if (participant != null) {
+          _onParticipantDeletedCtrl.add(participant);
+        }
         break;
       case 'typingStarted':
-        assert(participant != null);
-        _onTypingStartedCtrl.add(TypingEvent(this, participant));
+        if (participant != null) {
+          _onTypingStartedCtrl.add(TypingEvent(this, participant));
+        }
         break;
       case 'typingEnded':
-        assert(participant != null);
-        _onTypingEndedCtrl.add(TypingEvent(this, participant));
+        if (participant != null) {
+          _onTypingEndedCtrl.add(TypingEvent(this, participant));
+        }
         break;
       case 'synchronizationChanged':
         _onSynchronizationChangedCtrl.add(this);
