@@ -9,9 +9,7 @@ class ConversationUpdatedEvent {
   final Conversation conversation;
   final ConversationUpdateReason reason;
 
-  ConversationUpdatedEvent(this.conversation, this.reason)
-      : assert(conversation != null),
-        assert(reason != null);
+  ConversationUpdatedEvent(this.conversation, this.reason);
 }
 
 class UserUpdatedEvent {
@@ -19,9 +17,7 @@ class UserUpdatedEvent {
 
   final UserUpdateReason reason;
 
-  UserUpdatedEvent(this.user, this.reason)
-      : assert(user != null),
-        assert(reason != null);
+  UserUpdatedEvent(this.user, this.reason);
 }
 
 class NewMessageNotificationEvent {
@@ -30,61 +26,61 @@ class NewMessageNotificationEvent {
   final int messageIndex;
 
   NewMessageNotificationEvent(
-      this.conversationSid, this.messageSid, this.messageIndex)
-      : assert(conversationSid != null),
-        assert(messageSid != null),
-        assert(messageIndex != null);
+      this.conversationSid, this.messageSid, this.messageIndex);
 }
 
 class ConversationClient {
-  Conversations conversations;
+  Map<String, Conversation> conversations = <String, Conversation>{};
 
-  String myIdentity;
-  ConnectionState connectionState;
-  bool isReachabilityEnabled;
+  String? myIdentity;
+  ConnectionState connectionState = ConnectionState.UNKNOWN;
+  bool isReachabilityEnabled = false;
 
   /// Stream for the native client events.
-  StreamSubscription<dynamic> _clientStream;
+  late StreamSubscription<dynamic> _clientStream;
+
+  /// Stream for the native conversation events.
+  late StreamSubscription<dynamic> _conversationStream;
 
   /// Stream for the notification events.
-  StreamSubscription<dynamic> _notificationStream;
+  late StreamSubscription<dynamic> _notificationStream;
 
   final StreamController<bool> _onClientListenerAttachedCtrl =
       StreamController<bool>.broadcast();
 
   /// Called when client listener is listening and ready for client creation.
-  Stream<bool> onClientListenerAttached;
+  late Stream<bool> onClientListenerAttached;
 
   final StreamController<String> _onAddedToConversationNotificationCtrl =
       StreamController<String>.broadcast();
 
   /// Called when client receives a push notification for added to Conversation event.
-  Stream<String> onAddedToConversationNotification;
+  late Stream<String> onAddedToConversationNotification;
 
   final StreamController<Conversation> _onConversationAddedCtrl =
       StreamController<Conversation>.broadcast();
 
   /// Called when the current user has a conversation added to their conversation list, conversation status is not specified.
-  Stream<Conversation> onConversationAdded;
+  late Stream<Conversation> onConversationAdded;
 
   final StreamController<Conversation> _onConversationDeletedCtrl =
       StreamController<Conversation>.broadcast();
 
   /// Called when one of the conversation of the current user is deleted.
-  Stream<Conversation> onConversationDeleted;
+  late Stream<Conversation> onConversationDeleted;
 
   final StreamController<ConnectionState> _onConnectionStateCtrl =
       StreamController<ConnectionState>.broadcast();
 
   /// Called when client connnection state has changed.
-  Stream<ConnectionState> onConnectionState;
+  late Stream<ConnectionState> onConnectionState;
 
   final StreamController<ClientSynchronizationStatus>
       _onClientSynchronizationCtrl =
       StreamController<ClientSynchronizationStatus>.broadcast();
 
   /// Called when client synchronization status changes.
-  Stream<ClientSynchronizationStatus> onClientSynchronization;
+  late Stream<ClientSynchronizationStatus> onClientSynchronization;
 
   final StreamController<Conversation>
       _onConversationSynchronizationChangeCtrl =
@@ -93,7 +89,7 @@ class ConversationClient {
   /// Called when conversation synchronization status changed.
   ///
   /// Use [Conversation.synchronizationStatus] to obtain new conversation status.
-  Stream<Conversation> onConversationSynchronizationChange;
+  late Stream<Conversation> onConversationSynchronizationChange;
 
   final StreamController<ConversationUpdatedEvent> _onConversationUpdatedCtrl =
       StreamController<ConversationUpdatedEvent>.broadcast();
@@ -101,46 +97,46 @@ class ConversationClient {
   /// Called when the conversation is updated.
   ///
   /// [Conversation] synchronization updates are delivered via different callback.
-  Stream<ConversationUpdatedEvent> onConversationUpdated;
+  late Stream<ConversationUpdatedEvent> onConversationUpdated;
 
   final StreamController<ErrorInfo> _onErrorCtrl =
       StreamController<ErrorInfo>.broadcast();
 
   /// Called when an error condition occurs.
-  Stream<ErrorInfo> onError;
+  late Stream<ErrorInfo> onError;
 
   final StreamController<NewMessageNotificationEvent>
       _onNewMessageNotificationCtrl =
       StreamController<NewMessageNotificationEvent>.broadcast();
 
   /// Called when client receives a push notification for new message.
-  Stream<NewMessageNotificationEvent> onNewMessageNotification;
+  late Stream<NewMessageNotificationEvent> onNewMessageNotification;
 
   final StreamController<NotificationRegistrationEvent>
       _onNotificationRegisteredCtrl =
       StreamController<NotificationRegistrationEvent>.broadcast();
 
   /// Called when attempt to register device for notifications has completed.
-  Stream<NotificationRegistrationEvent> onNotificationRegistered;
+  late Stream<NotificationRegistrationEvent> onNotificationRegistered;
 
   final StreamController<String> _onRemovedFromConversationNotificationCtrl =
       StreamController<String>.broadcast();
 
   /// Called when client receives a push notification for removed from conversation event.
-  Stream<String> onRemovedFromConversationNotification;
+  late Stream<String> onRemovedFromConversationNotification;
 
   final StreamController<NotificationRegistrationEvent>
       _onNotificationDeregisteredCtrl =
       StreamController<NotificationRegistrationEvent>.broadcast();
 
   /// Called when attempt to register device for notifications has completed.
-  Stream<NotificationRegistrationEvent> onNotificationDeregistered;
+  late Stream<NotificationRegistrationEvent> onNotificationDeregistered;
 
   final StreamController<ErrorInfo> _onNotificationFailedCtrl =
       StreamController<ErrorInfo>.broadcast();
 
   /// Called when registering for push notifications fails.
-  Stream<ErrorInfo> onNotificationFailed;
+  late Stream<ErrorInfo> onNotificationFailed;
 
   //#region Token events
   final StreamController<void> _onTokenAboutToExpireCtrl =
@@ -149,7 +145,7 @@ class ConversationClient {
   /// Called when token is about to expire soon.
   ///
   /// In response, [ConversationClient] should generate a new token and call [ConversationClient.updateToken] as soon as possible.
-  Stream<void> onTokenAboutToExpire;
+  late Stream<void> onTokenAboutToExpire;
 
   final StreamController<void> _onTokenExpiredCtrl =
       StreamController<void>.broadcast();
@@ -157,25 +153,25 @@ class ConversationClient {
   /// Called when token has expired.
   ///
   /// In response, [ConversationClient] should generate a new token and call [ConversationClient.updateToken] as soon as possible.
-  Stream<void> onTokenExpired;
+  late Stream<void> onTokenExpired;
 
   final StreamController<User> _onUserSubscribedCtrl =
       StreamController<User>.broadcast();
 
   /// Called when a user is subscribed to and will receive realtime state updates.
-  Stream<User> onUserSubscribed;
+  late Stream<User> onUserSubscribed;
 
   final StreamController<User> _onUserUnsubscribedCtrl =
       StreamController<User>.broadcast();
 
   /// Called when a user is unsubscribed from and will not receive realtime state updates anymore.
-  Stream<User> onUserUnsubscribed;
+  late Stream<User> onUserUnsubscribed;
 
   final StreamController<UserUpdatedEvent> _onUserUpdatedCtrl =
       StreamController<UserUpdatedEvent>.broadcast();
 
   /// Called when user info is updated for currently loaded users.
-  Stream<UserUpdatedEvent> onUserUpdated;
+  late Stream<UserUpdatedEvent> onUserUpdated;
 
   ConversationClient() {
     onClientListenerAttached = _onClientListenerAttachedCtrl.stream;
@@ -201,10 +197,12 @@ class ConversationClient {
     onUserUnsubscribed = _onUserUnsubscribedCtrl.stream;
     onUserUpdated = _onUserUpdatedCtrl.stream;
 
-    conversations = Conversations();
     _clientStream = TwilioConversations.clientChannel
         .receiveBroadcastStream(0)
         .listen(_parseEvents);
+    _conversationStream = TwilioConversations.conversationChannel
+        .receiveBroadcastStream(0)
+        .listen(_parseConversationEvents);
     _notificationStream = TwilioConversations.notificationChannel
         .receiveBroadcastStream(0)
         .listen(_parseNotificationEvents);
@@ -212,9 +210,10 @@ class ConversationClient {
 
   void updateFromMap(Map<String, dynamic> json) {
     myIdentity = json['myIdentity'] as String;
-    connectionState = EnumToString.fromString<ConnectionState>(
-        ConnectionState.values, json['connectionState'] as String);
-    isReachabilityEnabled = json['isReachabilityEnabled'] as bool;
+    connectionState = EnumToString.fromString(
+            ConnectionState.values, json['connectionState']) ??
+        ConnectionState.UNKNOWN;
+    isReachabilityEnabled = json['isReachabilityEnabled'] ?? false;
   }
 
   /// Updates the authentication token for this client.
@@ -234,6 +233,7 @@ class ConversationClient {
   Future<void> shutdown() async {
     try {
       await _clientStream.cancel();
+      await _conversationStream.cancel();
       await _notificationStream.cancel();
       TwilioConversations.conversationClient = null;
       return await TwilioConversations.methodChannel
@@ -266,11 +266,61 @@ class ConversationClient {
     }
   }
 
+  //#region Conversations
+  Future<Conversation?> createConversation(
+      {required String friendlyName}) async {
+    final result = await TwilioConversations.methodChannel.invokeMethod(
+        'ConversationsMethods.createConversation',
+        {'friendlyName': friendlyName});
+    final conversationMap = Map<String, dynamic>.from(result);
+    updateConversationFromMap(conversationMap);
+    return conversations[conversationMap['sid']];
+  }
+
+  Future<Conversation?> getConversation(
+      {required String conversationSidOrUniqueName}) async {
+    final result = await TwilioConversations.methodChannel.invokeMethod(
+        'ConversationsMethods.getConversation',
+        {'conversationSidOrUniqueName': conversationSidOrUniqueName});
+    final conversationMap = Map<String, dynamic>.from(result);
+    updateConversationFromMap(conversationMap);
+    return conversations[conversationMap['sid']];
+  }
+
+  Future<List<Conversation>> getMyConversations() async {
+    final result = await TwilioConversations.methodChannel
+        .invokeMethod('ConversationsMethods.getMyConversations');
+    final conversationsMapList = result
+        .map((c) => Map<String, dynamic>.from(c))
+        .toList();
+
+    conversationsMapList.forEach((element) {
+      updateConversationFromMap(element);
+    });
+
+    return conversations.values.toList();
+  }
+
+  void updateConversationFromMap(Map<String, dynamic> map) {
+    var sid = map['sid'] as String;
+    if (conversations[sid] == null) {
+      conversations[sid] = Conversation.fromMap(map);
+    } else {
+      conversations[sid]?.updateFromMap(map);
+    }
+  }
+  //#endregion
+
   /// Parse native conversation client events to the right event streams.
   void _parseEvents(dynamic event) {
-    var eventMap = jsonDecode(event.toString());
-    final String eventName = eventMap['name'] as String;
-    final data = Map<String, dynamic>.from(eventMap['data'] as Map ?? {});
+    TwilioConversations.log(
+        'ConversationClient::parseEvents => received event: $event');
+    final eventName = event['name'] as String;
+
+    if (event['data'] == null) {
+      return;
+    }
+    final data = Map<String, dynamic>.from(event['data']);
 
     if (data['conversationClient'] != null) {
       final conversationClientMap =
@@ -278,26 +328,27 @@ class ConversationClient {
       updateFromMap(conversationClientMap);
     }
 
-    ErrorInfo exception;
-    if (eventMap['error'] != null) {
+    ErrorInfo? exception;
+    if (event['error'] != null) {
       final errorMap =
-          Map<String, dynamic>.from(eventMap['error'] as Map<dynamic, dynamic>);
+          Map<String, dynamic>.from(event['error'] as Map<dynamic, dynamic>);
       exception = ErrorInfo(errorMap['code'] as int,
           errorMap['message'] as String, errorMap['status'] as int);
     }
 
-    var conversationSid = data['conversationSid'] as String;
+    var conversationSid = data['conversationSid'] as String?;
 
-    Map<String, dynamic> conversationMap;
+    Map<String, dynamic>? conversationMap;
     if (data['conversation'] != null) {
       conversationMap = Map<String, dynamic>.from(
           data['conversation'] as Map<dynamic, dynamic>);
       conversationSid = conversationMap['sid'] as String;
     }
 
-    Map<String, dynamic> userMap;
+    Map<String, dynamic>? userMap;
     if (data['user'] != null) {
-      // userMap = Map<String, dynamic>.from(data['user'] as Map<dynamic, dynamic>);
+      userMap =
+          Map<String, dynamic>.from(data['user'] as Map<dynamic, dynamic>);
     }
     dynamic reason;
     if (data['reason'] != null) {
@@ -317,74 +368,93 @@ class ConversationClient {
         _onClientListenerAttachedCtrl.add(true);
         break;
       case 'addedToConversationNotification':
-        assert(conversationSid != null);
-        _onAddedToConversationNotificationCtrl.add(conversationSid);
+        if (conversationSid != null) {
+          _onAddedToConversationNotificationCtrl.add(conversationSid);
+        }
         break;
       case 'conversationAdded':
-        assert(conversationMap != null);
-        assert(conversationSid != null);
-        Conversations.updateConversationFromMap(conversationMap);
-        _onConversationAddedCtrl
-            .add(Conversations.conversationsMap[conversationSid]);
+        if (conversationMap != null && conversationSid != null) {
+          updateConversationFromMap(conversationMap);
+          final conversation = conversations[conversationSid];
+          if (conversation != null) {
+            _onConversationAddedCtrl.add(conversation);
+          }
+        }
         break;
       case 'conversationDeleted':
-        assert(conversationMap != null);
-        assert(conversationSid != null);
-        final conversation = Conversations.conversationsMap[conversationSid];
-        Conversations.conversationsMap.remove(conversationSid);
-        _onConversationDeletedCtrl.add(conversation);
+        if (conversationMap != null && conversationSid != null) {
+          final conversation = conversations[conversationSid];
+          conversations.remove(conversationSid);
+          if (conversation != null) {
+            _onConversationDeletedCtrl.add(conversation);
+          }
+        }
         break;
       case 'conversationSynchronizationChange':
-        assert(conversationMap != null);
-        assert(conversationSid != null);
-        Conversations.updateConversationFromMap(conversationMap);
-        _onConversationSynchronizationChangeCtrl
-            .add(Conversations.conversationsMap[conversationSid]);
+        if (conversationMap != null && conversationSid != null) {
+          updateConversationFromMap(conversationMap);
+          final conversation = conversations[conversationSid];
+          if (conversation != null) {
+            _onConversationSynchronizationChangeCtrl.add(conversation);
+          }
+        }
         break;
       case 'conversationUpdated':
-        assert(conversationMap != null);
-        assert(reason != null);
-        assert(conversationSid != null);
-        Conversations.updateConversationFromMap(conversationMap);
-        _onConversationUpdatedCtrl.add(ConversationUpdatedEvent(
-          Conversations.conversationsMap[conversationSid],
-          reason as ConversationUpdateReason,
-        ));
+        if (conversationMap != null &&
+            reason != null &&
+            conversationSid != null) {
+          final conversation = conversations[conversationSid];
+          if (conversation != null) {
+            updateConversationFromMap(conversationMap);
+            _onConversationUpdatedCtrl.add(ConversationUpdatedEvent(
+              conversation,
+              reason as ConversationUpdateReason,
+            ));
+          }
+        }
         break;
       case 'clientSynchronization':
         var synchronizationStatus = EnumToString.fromString(
-            ClientSynchronizationStatus.values,
-            data['synchronizationStatus'] as String);
-        assert(synchronizationStatus != null);
+                ClientSynchronizationStatus.values,
+                data['synchronizationStatus'] as String) ??
+            ClientSynchronizationStatus.UNKNOWN;
         _onClientSynchronizationCtrl.add(synchronizationStatus);
         break;
       case 'connectionStateChange':
         var newConnectionState = EnumToString.fromString(
             ConnectionState.values, data['connectionState'] as String);
-        assert(newConnectionState != null);
-        connectionState = newConnectionState;
-        _onConnectionStateCtrl.add(newConnectionState);
+        if (newConnectionState != null) {
+          connectionState = newConnectionState;
+          _onConnectionStateCtrl.add(newConnectionState);
+        } else {
+          TwilioConversations.log(
+              "ConversationClient => case 'connectionStateChange' => Attempting to operate on NULL.");
+        }
         break;
       case 'error':
-        assert(exception != null);
-        _onErrorCtrl.add(exception);
+        if (exception != null) {
+          _onErrorCtrl.add(exception);
+        }
         break;
       case 'newMessageNotification':
-        var messageSid = data['messageSid'] as String;
-        var messageIndex = data['messageIndex'] as int;
-        assert(conversationSid != null);
-        assert(messageSid != null);
-        assert(messageIndex != null);
-        _onNewMessageNotificationCtrl.add(NewMessageNotificationEvent(
-            conversationSid, messageSid, messageIndex));
+        var messageSid = data['messageSid'] as String?;
+        var messageIndex = data['messageIndex'] as int?;
+        if (conversationSid != null &&
+            messageSid != null &&
+            messageIndex != null) {
+          _onNewMessageNotificationCtrl.add(NewMessageNotificationEvent(
+              conversationSid, messageSid, messageIndex));
+        }
         break;
       case 'notificationFailed':
-        assert(exception != null);
-        _onNotificationFailedCtrl.add(exception);
+        if (exception != null) {
+          _onNotificationFailedCtrl.add(exception);
+        }
         break;
       case 'removedFromConversationNotification':
-        assert(conversationSid != null);
-        _onRemovedFromConversationNotificationCtrl.add(conversationSid);
+        if (conversationSid != null) {
+          _onRemovedFromConversationNotificationCtrl.add(conversationSid);
+        }
         break;
       case 'tokenAboutToExpire':
         _onTokenAboutToExpireCtrl.add(null);
@@ -400,41 +470,52 @@ class ConversationClient {
         //_onUserSubscribedCtrl.add(User.fromJson(userMap));
         break;
       case 'userUnsubscribed':
-        //TODO Fix
-        // assert(userMap != null);
+        if (userMap == null) {
+          TwilioConversations.log(
+              'ConversationClient => case \'userUnsubscribed\' => userMap is NULL.');
+          return;
+        }
         // var user = users.getUserById(userMap['identity']);
         // user._updateFromMap(userMap);
         // users.subscribedUsers.removeWhere((u) => u.identity == userMap['identity']);
-        _onUserUnsubscribedCtrl.add(User.fromJson(userMap));
+        _onUserUnsubscribedCtrl.add(User.fromMap(userMap));
         break;
       case 'userUpdated':
-        assert(userMap != null);
-        assert(reason != null);
-        // users._updateFromMap({
-        //   'subscribedUsers': [userMap]
-        // });
-        _onUserUpdatedCtrl.add(UserUpdatedEvent(
-            User.fromJson(userMap), reason as UserUpdateReason));
+        if (userMap != null && reason != null) {
+          // users._updateFromMap({
+          //   'subscribedUsers': [userMap]
+          // });
+          _onUserUpdatedCtrl.add(UserUpdatedEvent(
+              User.fromMap(userMap), reason as UserUpdateReason));
+        }
         break;
       default:
-        TwilioConversations.log("Event '$eventName' not yet implemented");
+        TwilioConversations.log('Event \'$eventName\' not yet implemented');
         break;
+    }
+  }
+
+  void _parseConversationEvents(dynamic event) {
+    final data = Map<String, dynamic>.from(event['data']);
+    final String? conversationSid = data['conversationSid'];
+    if (conversationSid != null) {
+      conversations[conversationSid]?.parseEvents(event);
     }
   }
 
   /// Parse notification events to the right event streams.
   void _parseNotificationEvents(dynamic event) {
-    final jsonMap = jsonDecode(event.toString()) as Map<String, dynamic>;
-    final eventName = jsonMap['name'] as String;
+    final eventMap = Map<String, dynamic>.from(event);
+    final eventName = eventMap['name'] as String;
     TwilioConversations.log(
-        "ConversationClient => Event '$eventName' => ${jsonMap["data"]}, error: ${jsonMap["error"]}");
+        'ConversationClient => Event \'$eventName\' => ${eventMap['data']}, error: ${eventMap['error']}');
     final data =
-        Map<String, dynamic>.from(jsonMap['data'] as Map<String, dynamic>);
+        Map<String, dynamic>.from(eventMap['data'] as Map<String, dynamic>);
 
-    ErrorInfo exception;
-    if (jsonMap['error'] != null) {
+    ErrorInfo? exception;
+    if (eventMap['error'] != null) {
       final errorMap =
-          Map<String, dynamic>.from(jsonMap['error'] as Map<dynamic, dynamic>);
+          Map<String, dynamic>.from(eventMap['error'] as Map<dynamic, dynamic>);
       exception = ErrorInfo(errorMap['code'] as int,
           errorMap['message'] as String, errorMap['status'] as int);
     }
@@ -450,7 +531,7 @@ class ConversationClient {
         break;
       default:
         TwilioConversations.log(
-            "Notification event '$eventName' not yet implemented");
+            'Notification event \'$eventName\' not yet implemented');
         break;
     }
   }
