@@ -1,2 +1,8 @@
-export 'parts/parts.dart';
 export 'twilio_conversations.dart';
+export 'attributes.dart';
+export 'error_info.dart';
+export 'conversation/src.dart';
+export 'conversation_client/src.dart';
+export 'message/src.dart';
+export 'participant/src.dart';
+export 'user/src.dart';
